Refactor withdraw DataViews: extract shared fields, fix double modal, migrate Pending Requests table
- Extract shared field definitions, types, and constants into withdraw-fields.tsx
  to eliminate duplication between WithdrawRequests and PendingRequestsTable
- Fix getItemId to return string (DataViews expects string, not number)
- Fix double confirmation modal by removing DokanModal and using DataViews
  built-in isDestructive confirmation instead
- Migrate Pending Requests table in PaymentDetails to new PendingRequestsTable
  component using the same DataViews pattern
- Fix implicit any type on onViewChange, simplify cancelRequestId to number|null
- Fix stale closure in withdrawRequestsCompat useMemo dependency array
- Delete unused RequestList.tsx (replaced by PendingRequestsTable)
