+++ src/dashboard/withdraw/WithdrawRequests.tsx
@@ -3,29 +3,18 @@ import { useEffect, useState, useMemo, useCallback } from '@wordpress/element';
 import { useToast } from '@getdokan/dokan-ui';
 import apiFetch from '@wordpress/api-fetch';
 import { addQueryArgs } from '@wordpress/url';
-import { DataViews, DokanModal } from '@dokan/components';
-import PriceHtml from '../../components/PriceHtml';
-import DateTimeHtml from '../../components/DateTimeHtml';
+import { DataViews } from '@dokan/components';
 import { useWithdraw } from './Hooks/useWithdraw';
 import RequestWithdrawBtn from './RequestWithdrawBtn';
 import { useWithdrawSettings } from './Hooks/useWithdrawSettings';
 import { useCurrentUser } from '@dokan/hooks';
 import { useBalance } from './Hooks/useBalance';
-
-type WithdrawStatus = 'pending' | 'approved' | 'cancelled';
-
-interface WithdrawRequest {
-    id: number;
-    user_id: number;
-    amount: number;
-    status: string;
-    method: string;
-    method_title: string;
-    created: string;
-    charge: number;
-    receivable: number;
-    note: string;
-}
+import {
+    DEFAULT_LAYOUTS,
+    getFieldsForStatus,
+    type WithdrawRequest,
+    type WithdrawStatus,
+} from './withdraw-fields';
 
 interface WithdrawSummary {
     total: number;
@@ -34,16 +23,14 @@ interface WithdrawSummary {
     cancelled: number;
 }
 
-const DEFAULT_LAYOUTS = {
-    table: { density: 'comfortable' },
-    list: {},
-};
-
-const allStatusLabels = {
-    pending: __( 'Pending Review', 'dokan-lite' ),
-    approved: __( 'Approved', 'dokan-lite' ),
-    cancelled: __( 'Cancelled', 'dokan-lite' ),
-};
+interface WithdrawView {
+    perPage: number;
+    page: number;
+    search: string;
+    type: string;
+    status: WithdrawStatus;
+    fields: string[];
+}
 
 function WithdrawRequests() {
     const withdrawSettings = useWithdrawSettings();
@@ -56,8 +43,6 @@ function WithdrawRequests() {
     const [ isLoading, setIsLoading ] = useState( true );
     const [ totalItems, setTotalItems ] = useState( 0 );
     const [ totalPages, setTotalPages ] = useState( 0 );
-    const [ isOpen, setIsOpen ] = useState( false );
-    const [ cancelRequestId, setCancelRequestId ] = useState( '' );
     const [ summary, setSummary ] = useState< WithdrawSummary >( {
         total: 0,
         pending: 0,
@@ -65,211 +50,17 @@ function WithdrawRequests() {
         cancelled: 0,
     } );
 
-    // Stub withdraw requests object for RequestWithdrawBtn compatibility
-    const withdrawRequestsCompat = {
-        data,
-        isLoading,
-        refresh: () => fetchWithdrawRequests(),
-        totalItems,
-        totalPages,
-        view: null,
-        setView: () => {},
-        setData: () => {},
-        error: null,
-        fetchWithdrawRequests: () => {},
-        lastPayload: null,
-    };
-
-    const pendingFields = [
-        {
-            id: 'amount',
-            label: __( 'Amount', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <PriceHtml price={ item?.amount } />
-            ),
-        },
-        {
-            id: 'method_title',
-            label: __( 'Method', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <span>{ item?.method_title }</span>
-            ),
-        },
-        {
-            id: 'created',
-            label: __( 'Date', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <DateTimeHtml date={ item?.created } />
-            ),
-        },
-        {
-            id: 'charge',
-            label: __( 'Charge', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <PriceHtml price={ item?.charge } />
-            ),
-        },
-        {
-            id: 'receivable',
-            label: __( 'Receivable', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <PriceHtml price={ item?.receivable } />
-            ),
-        },
-        {
-            id: 'status',
-            label: __( 'Status', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <span>
-                    { allStatusLabels[ item?.status as WithdrawStatus ] }
-                </span>
-            ),
-        },
-    ];
-
-    const approvedFields = [
-        {
-            id: 'amount',
-            label: __( 'Amount', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <PriceHtml price={ item?.amount } />
-            ),
-        },
-        {
-            id: 'method_title',
-            label: __( 'Method', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <span>{ item?.method_title }</span>
-            ),
-        },
-        {
-            id: 'created',
-            label: __( 'Date', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <DateTimeHtml date={ item?.created } />
-            ),
-        },
-        {
-            id: 'charge',
-            label: __( 'Charge', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <PriceHtml price={ item?.charge } />
-            ),
-        },
-        {
-            id: 'receivable',
-            label: __( 'Receivable', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <PriceHtml price={ item?.receivable } />
-            ),
-        },
-    ];
-
-    const cancelledFields = [
-        {
-            id: 'amount',
-            label: __( 'Amount', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <PriceHtml price={ item?.amount } />
-            ),
-        },
-        {
-            id: 'method_title',
-            label: __( 'Method', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <span>{ item?.method_title }</span>
-            ),
-        },
-        {
-            id: 'created',
-            label: __( 'Date', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <DateTimeHtml date={ item?.created } />
-            ),
-        },
-        {
-            id: 'charge',
-            label: __( 'Charge', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <PriceHtml price={ item?.charge } />
-            ),
-        },
-        {
-            id: 'receivable',
-            label: __( 'Receivable', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <PriceHtml price={ item?.receivable } />
-            ),
-        },
-        {
-            id: 'note',
-            label: __( 'Note', 'dokan-lite' ),
-            enableSorting: false,
-            render: ( { item }: { item: WithdrawRequest } ) => (
-                <span>{ item?.note }</span>
-            ),
-        },
-    ];
-
-    const getFieldsForStatus = ( status: WithdrawStatus ) => {
-        switch ( status ) {
-            case 'pending':
-                return pendingFields;
-            case 'approved':
-                return approvedFields;
-            case 'cancelled':
-                return cancelledFields;
-            default:
-                return pendingFields;
-        }
-    };
-
-    const [ view, setView ] = useState( {
+    const [ view, setView ] = useState< WithdrawView >( {
         perPage: 10,
         page: 1,
         search: '',
         type: 'table',
-        status: 'pending' as WithdrawStatus,
-        fields: pendingFields.map( ( field ) => field.id ),
+        status: 'pending',
+        fields: getFieldsForStatus( 'pending' ).map( ( field ) => field.id ),
     } );
 
     const currentStatus = view.status;
     const fields = getFieldsForStatus( currentStatus );
-
-    const actions = useMemo( () => {
-        if ( currentStatus !== 'pending' ) {
-            return [];
-        }
-
-        return [
-            {
-                id: 'withdraw-cancel',
-                isEligible: () => true,
-                label: () => __( 'Cancel', 'dokan-lite' ),
-                isDestructive: true,
-                callback: ( [ item ]: WithdrawRequest[] ) => {
-                    setCancelRequestId( String( item.id ) );
-                    setIsOpen( true );
-                },
-            },
-        ];
-    }, [ currentStatus ] );
 
     const fetchSummary = useCallback( async () => {
         try {
@@ -324,29 +115,66 @@ function WithdrawRequests() {
         }
     }, [ view.perPage, view.page, view.status, currentUser?.id ] );
 
-    const cancelPendingRequest = useCallback( () => {
-        withdrawHook
-            .updateWithdraw( Number( cancelRequestId ), {
-                status: 'cancelled',
-            } )
-            .then( () => {
-                toast( {
-                    type: 'success',
-                    title: __( 'Request cancelled successfully', 'dokan-lite' ),
-                } );
-                void fetchWithdrawRequests();
-                void fetchSummary();
-            } )
-            .catch( () => {
-                toast( {
-                    type: 'error',
-                    title: __( 'Failed to cancel request', 'dokan-lite' ),
-                } );
-            } )
-            .finally( () => {
-                setIsOpen( false );
-            } );
-    }, [ cancelRequestId, withdrawHook, fetchWithdrawRequests, fetchSummary ] );
+    // Stub withdraw requests object for RequestWithdrawBtn compatibility
+    // TODO: Refactor RequestWithdrawBtn to accept a simpler interface
+    // instead of the full UseWithdrawRequestsReturn shape.
+    const withdrawRequestsCompat = useMemo(
+        () => ( {
+            data,
+            isLoading,
+            refresh: () => fetchWithdrawRequests(),
+            totalItems,
+            totalPages,
+            view: { perPage: 10, page: 1, search: '', type: 'table', titleField: 'amount' },
+            setView: () => {},
+            setData: () => {},
+            error: null,
+            fetchWithdrawRequests: () => {},
+            lastPayload: null,
+        } as any ),
+        [ data, isLoading, totalItems, totalPages, fetchWithdrawRequests ]
+    );
+
+    const actions = useMemo( () => {
+        if ( currentStatus !== 'pending' ) {
+            return [];
+        }
+
+        return [
+            {
+                id: 'withdraw-cancel',
+                isEligible: () => true,
+                label: () => __( 'Cancel', 'dokan-lite' ),
+                isDestructive: true,
+                callback: ( [ item ]: WithdrawRequest[] ) => {
+                    withdrawHook
+                        .updateWithdraw( item.id, {
+                            status: 'cancelled',
+                        } )
+                        .then( () => {
+                            toast( {
+                                type: 'success',
+                                title: __(
+                                    'Request cancelled successfully',
+                                    'dokan-lite'
+                                ),
+                            } );
+                            void fetchWithdrawRequests();
+                            void fetchSummary();
+                        } )
+                        .catch( () => {
+                            toast( {
+                                type: 'error',
+                                title: __(
+                                    'Failed to cancel request',
+                                    'dokan-lite'
+                                ),
+                            } );
+                        } );
+                },
+            },
+        ];
+    }, [ currentStatus, withdrawHook, fetchWithdrawRequests, fetchSummary ] );
 
     useEffect( () => {
         void fetchWithdrawRequests();
@@ -393,7 +221,7 @@ function WithdrawRequests() {
         totalPages,
     };
 
-    const onViewChange = useCallback( ( newView ) => {
+    const onViewChange = useCallback( ( newView: typeof view ) => {
         setView( newView );
     }, [] );
 
@@ -410,7 +238,7 @@ function WithdrawRequests() {
                 data={ data }
                 defaultLayouts={ DEFAULT_LAYOUTS }
                 fields={ fields }
-                getItemId={ ( item: WithdrawRequest ) => item.id }
+                getItemId={ ( item: WithdrawRequest ) => String( item.id ) }
                 onChangeView={ onViewChange }
                 paginationInfo={ paginationInfo }
                 view={ view }
@@ -420,24 +248,6 @@ function WithdrawRequests() {
                 tabs={ tabs }
             />
 
-            <DokanModal
-                isOpen={ isOpen }
-                namespace="cancel-request-confirmation"
-                onConfirm={ cancelPendingRequest }
-                onClose={ () => setIsOpen( false ) }
-                dialogTitle={ __( 'Cancel Withdraw Request', 'dokan-lite' ) }
-                confirmationTitle={ __(
-                    'Are you sure you want to proceed?',
-                    'dokan-lite'
-                ) }
-                confirmationDescription={ __(
-                    'Do you want to proceed for cancelling the withdraw request?',
-                    'dokan-lite'
-                ) }
-                confirmButtonText={ __( 'Yes, Cancel', 'dokan-lite' ) }
-                cancelButtonText={ __( 'Close', 'dokan-lite' ) }
-                loading={ withdrawHook.isLoading }
-            />
         </div>
     );
 }
